feat: nvidia driver exposed in docker to enable opengl
opengl is important for rqt* and other graphical
diagnostic tools
as well as simulators or using machine learning on gpus.

This image is specific to nvidia-gpus, but allows to use
the nvidia gpu in the docker image.

clarified README.md

feat: enabled nvidia-host specific container for opengl

feat: added ros2-lgsvl-bridge

The readme was assuming the opt-in for nvidia-docker
Now the readme assumes that there is a branch
for the host-specific foxy_nvidia